Fully untangle
