+++ packages/core/package.json
@@ -11,7 +11,8 @@
     "ts-loader": "^9.4.4",
     "ts-mocha": "^10.0.0",
     "chai": "^4.4.1",
-    "@types/chai": "^4.3.16"
+    "@types/chai": "^4.3.16",
+    "@types/jsdom": "^21.1.3"
   },
   "exports": {
     "./*": "./src/*.ts"
