+++ packages/frontend/src/scripts/components/no_settings.ts
@@ -1,0 +1,15 @@
+import {Simulation} from "@xivgear/core/sims/sim_types";
+
+/**
+ * Basic implementation of {@link Simulation.makeResultDisplay} for sims which do not actually
+ * have any settings.
+ */
+export function noSimSettings() {
+    const outerDiv = document.createElement("div");
+    const header = document.createElement("h1");
+    header.textContent = "No Settings for This Simulation";
+    outerDiv.replaceChildren(header);
+    return outerDiv;
+}
+
+
